requirements.txt
fastapi
uvicorn
torch
transformers
sentence-transformers
faiss-cpu
beautifulsoup4
requests
numpy